Split Query into StructuredQuery and NativeQuery
