+++ frontend/src/metabase-lib/lib/NativeQuery.js
@@ -1,0 +1,7 @@
+import Query from "./Query";
+
+export default class NativeQuery extends Query {
+    isNative() {
+        return true;
+    }
+}
